+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -9,6 +9,18 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.1.1.6
+-------
+Changed
+*******
+- Catch OperationCanceledException instead of TaskCanceledException to avoid useless server log spams when FindGame is canceled by clients.
+
+4.1.1.5
+-------
+Changed
+*******
+- Add Reason in JoiningPartyContext
+
 4.1.1.4
 -------
 Fixed